[0/5] Supress implicit conversion warning.
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@3926 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ src/runtime/components/component_type.cpp
@@ -48,8 +48,10 @@ namespace hpx { namespace components
             result += "[" + boost::lexical_cast<std::string>(type) + "]";
         else {
             result += "[" + 
-                boost::lexical_cast<std::string>(get_derived_type(type)) + 
-                "(" + boost::lexical_cast<std::string>(get_base_type(type)) + ")"
+                boost::lexical_cast<std::string>
+                  (static_cast<int>(get_derived_type(type))) + 
+                "(" + boost::lexical_cast<std::string>
+                    (static_cast<int>(get_base_type(type))) + ")"
                 "]";
         }
         return result;
